# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "meap api document",
        "description": "meap api document",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:8000"
        }
    ],
    "paths": {
...
        "/meal-plans": {
            "get": {
                "tags": [
                    "MealPlans"
                ],
                "summary": "献立一覧を取得",
                "description": "指定した期間の献立一覧を取得します。date_from と date_to は必須のクエリパラメータです。",
                "operationId": "97ebdfd313f0459c06cf0a048ab8b1d3",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MealPlanIndexDateFromParam"
                    },
                    {
                        "$ref": "#/components/parameters/MealPlanIndexDateToParam"
                    },
                    {
                        "$ref": "#/components/parameters/MealPlanIndexIncludeIngredientsParam"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MealPlanIndexSuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrors"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "MealPlans"
                ],
                "summary": "献立を作成",
                "description": "献立を作成します。",
                "operationId": "8757b0ba77d78ebcf68636c44897a11c",
                "requestBody": {
                    "$ref": "#/components/requestBodies/MealPlanStoreRequest"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MealPlanStoreSuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/meal-plans/{date}": {
            "get": {
                "tags": [
                    "MealPlans"
                ],
                "summary": "献立の詳細を取得",
                "description": "指定した日付（Y-m-d）の献立を1件取得します。",
                "operationId": "361429cc34688cb6836990e2b072d43d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MealPlanDateParam"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MealPlanShowSuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/meal-plans/{id}": {
            "put": {
                "tags": [
                    "MealPlans"
                ],
                "summary": "献立を更新",
                "operationId": "c12aa611abd0e80f6f02214b178d3674",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MealPlanIdParam"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/MealPlanUpdateRequest"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MealPlanUpdateSuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MealPlans"
                ],
                "summary": "献立を削除",
                "operationId": "0492908ed27eeea9284a4d7d9522425f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MealPlanIdParam"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MealPlanDestroySuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/meal-plans/{mealPlanId}/meals/{mealId}": {
            "delete": {
                "tags": [
                    "MealPlans"
                ],
                "summary": "献立の1食を削除",
                "operationId": "b792b9051c27e46abb92d04576fe4c78",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MealPlanIdPathParam"
                    },
                    {
                        "$ref": "#/components/parameters/MealIdParam"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MealPlanMealDestroySuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BaseApiResponse": {
                "description": "成功（success, message, data: null）",
                "required": [
                    "success",
                    "message",
                    "data"
                ],
                "properties": {
                    "success": {
                        "description": "成功フラグ",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "結果メッセージ",
                        "type": "string",
                        "example": "操作が完了しました。"
                    },
                    "data": {
                        "description": "store/update/destroy 等では常に null",
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "BaseApiIndexResponse": {
                "description": "成功（success, message, data: 配列, total）",
                "required": [
                    "success",
                    "message",
                    "data",
                    "total"
                ],
                "properties": {
                    "success": {
                        "description": "成功フラグ",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "結果メッセージ",
                        "type": "string",
                        "example": "一覧を取得しました。"
                    },
                    "data": {
                        "description": "一覧データ",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "total": {
                        "description": "総件数",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "MealPlanStoreRequest": {
                "required": [
                    "date",
                    "meals"
                ],
                "properties": {
                    "date": {
                        "description": "献立の日付（Y-m-d）",
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-03"
                    },
                    "meals": {
                        "description": "献立メニュー",
                        "type": "array",
                        "items": {
                            "required": [
                                "categoryId",
                                "recipes",
                                "order"
                            ],
                            "properties": {
                                "id": {
                                    "description": "メニューID（新規の場合は不要）",
                                    "type": "string",
                                    "format": "uuid",
                                    "example": "a0fbbf74-1816-406e-99b7-7ef1e3e365f4",
                                    "nullable": true
                                },
                                "categoryId": {
                                    "description": "献立カテゴリID",
                                    "type": "string",
                                    "format": "uuid",
                                    "example": "a0fbbf74-1816-406e-99b7-7ef1e3e365f4"
                                },
                                "order": {
                                    "description": "表示順（0以上）",
                                    "type": "integer",
                                    "example": 0
                                },
                                "recipes": {
                                    "description": "料理一覧（1件以上必須）。各要素の order が1食内の並び順（0始まり）",
                                    "type": "array",
                                    "items": {
                                        "required": [
                                            "id",
                                            "order"
                                        ],
                                        "properties": {
                                            "id": {
                                                "description": "レシピID",
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "a0fbbf74-1816-406e-99b7-7ef1e3e365f4"
                                            },
                                            "order": {
                                                "description": "1食内の並び順（0以上）",
                                                "type": "integer",
                                                "example": 0
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "minItems": 1
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "MealPlanUpdateRequest": {
                "required": [
                    "meals"
                ],
                "properties": {
                    "meals": {
                        "description": "献立メニュー",
                        "type": "array",
                        "items": {
                            "required": [
                                "categoryId",
                                "recipes",
                                "order"
                            ],
                            "properties": {
                                "id": {
                                    "description": "メニューID（新規の場合は不要）",
                                    "type": "string",
                                    "format": "uuid",
                                    "example": "a0fbbf74-1816-406e-99b7-7ef1e3e365f4",
                                    "nullable": true
                                },
                                "categoryId": {
                                    "description": "献立カテゴリID",
                                    "type": "string",
                                    "format": "uuid",
                                    "example": "a0fbbf74-1816-406e-99b7-7ef1e3e365f4"
                                },
                                "order": {
                                    "description": "表示順（0以上）",
                                    "type": "integer",
                                    "example": 0
                                },
                                "recipes": {
                                    "description": "料理一覧（1件以上必須）。各要素の order が1食内の並び順（0始まり）",
                                    "type": "array",
                                    "items": {
                                        "required": [
                                            "id",
                                            "order"
                                        ],
                                        "properties": {
                                            "id": {
                                                "description": "レシピID",
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "a0fbbf74-1816-406e-99b7-7ef1e3e365f4"
                                            },
                                            "order": {
                                                "description": "1食内の並び順（0以上）",
                                                "type": "integer",
                                                "example": 0
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "minItems": 1
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "MealPlanIndexResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseApiIndexResponse"
                    },
                    {
                        "required": [
                            "data"
                        ],
                        "properties": {
                            "data": {
                                "description": "日付と献立メニュー",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/MealPlan"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "MealPlanShowResponse": {
                "required": [
                    "success",
                    "message",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "献立(2024-01-15)を取得しました。"
                    },
                    "data": {
                        "$ref": "#/components/schemas/MealPlan"
                    }
                },
                "type": "object"
            },
...
            "MealPlan": {
                "required": [
                    "date",
                    "meals"
                ],
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "string",
                        "example": "1"
                    },
                    "date": {
                        "description": "日付",
                        "type": "string",
                        "format": "date",
                        "example": "2023-10-05"
                    },
                    "meals": {
                        "description": "献立メニュー",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MealPlanItem"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "Unauthorized": {
                "description": "認証エラー",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "認証が必要です。"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "リソースが見つかりませんでした",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "指定されたレコードが見つかりません。"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationErrors": {
                "description": "バリデーションエラー",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "入力内容に誤りがあります。"
                                },
                                "errors": {
                                    "properties": {
                                        "name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "名前は必須です。"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "MealPlanIndexSuccess": {
                "description": "献立を5件取得しました。",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MealPlanIndexResponse"
                        }
                    }
                }
            },
            "MealPlanStoreSuccess": {
                "description": "献立(2024-01-15)を作成しました。",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BaseApiResponse"
                        }
                    }
                }
            },
            "MealPlanShowSuccess": {
                "description": "献立(2024-01-15)を取得しました。",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MealPlanShowResponse"
                        }
                    }
                }
            },
            "MealPlanUpdateSuccess": {
                "description": "献立(2024-01-15)を更新しました。",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BaseApiResponse"
                        }
                    }
                }
            },
            "MealPlanDestroySuccess": {
                "description": "献立(2024-01-15)を削除しました。",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BaseApiResponse"
                        }
                    }
                }
            },
            "MealPlanMealDestroySuccess": {
                "description": "献立の1食を削除しました。",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BaseApiResponse"
                        }
                    }
                }
            },
...
        },
        "parameters": {
...
            "MealPlanIndexDateFromParam": {
                "name": "date_from",
                "in": "query",
                "description": "取得対象の開始日",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-01-01"
                }
            },
            "MealPlanIndexDateToParam": {
                "name": "date_to",
                "in": "query",
                "description": "取得対象の終了日",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-01-31"
                }
            },
            "MealPlanIndexIncludeIngredientsParam": {
                "name": "include_ingredients",
                "in": "query",
                "description": "レスポンスにレシピの食材一覧を含めるかどうか",
                "required": false,
                "schema": {
                    "type": "boolean",
                    "example": true
                }
            },
            "MealPlanDateParam": {
                "name": "date",
                "in": "path",
                "description": "献立の日付（Y-m-d）",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-01-15"
                }
            },
            "MealPlanIdParam": {
                "name": "id",
                "in": "path",
                "description": "献立ID",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "1"
                }
            },
            "MealPlanIdPathParam": {
                "name": "mealPlanId",
                "in": "path",
                "description": "献立ID",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "uuid"
                }
            },
            "MealIdParam": {
                "name": "mealId",
                "in": "path",
                "description": "献立の1食ID",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "uuid"
                }
            },
...
        },
        "requestBodies": {
...
            "MealPlanStoreRequest": {
                "description": "献立作成。date は必須。meals は1件以上必須。",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MealPlanStoreRequest"
                        }
                    }
                }
            },
            "MealPlanUpdateRequest": {
                "description": "献立更新。meals は1件以上必須。",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MealPlanUpdateRequest"
                        }
                    }
                }
            },
...
        }
    },
...
}